기본 요구 사항 구현 완료ㅕ
