+++ settings.gradle
@@ -1,2 +1,1 @@
-rootProject.name = '0-sprint-mission'
-
+rootProject.name = '2-sprint-mission'
