+++ src/main/java/com/sprint/mission/discodeit/JavaApplication.java
@@ -1,42 +1,16 @@
 package com.sprint.mission.discodeit;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
-import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.repository.ChannelRepository;
-import com.sprint.mission.discodeit.repository.MessageRepository;
-import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.repository.file.FileChannelRepository;
-import com.sprint.mission.discodeit.repository.file.FileMessageRepository;
-import com.sprint.mission.discodeit.repository.file.FileUserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.service.UserService;
-import com.sprint.mission.discodeit.service.basic.BasicChannelService;
-import com.sprint.mission.discodeit.service.basic.BasicMessageService;
-import com.sprint.mission.discodeit.service.basic.BasicUserService;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
 public class JavaApplication {
-    // test
-    static User setupUser(UserService userService) {
-        User user = userService.create("woody", "woody@codeit.com", "woody1234");
-        return user;
-    }
-
-    static Channel setupChannel(ChannelService channelService) {
-        Channel channel = channelService.create(ChannelType.PUBLIC, "공지", "공지 채널입니다.");
-        return channel;
-    }
-
-    static void messageCreateTest(MessageService messageService, Channel channel, User author) {
-        Message message = messageService.create("안녕하세요.", channel.getId(), author.getId());
-        System.out.println("메시지 생성: " + message.getId());
-    }
-
-
     public static void main(String[] args) {
-        // 레포지토리 초기화
+        /*// 레포지토리 초기화
         UserRepository userRepository = new FileUserRepository();
         ChannelRepository channelRepository = new FileChannelRepository();
         MessageRepository messageRepository = new FileMessageRepository();
@@ -45,11 +19,24 @@ public class JavaApplication {
         UserService userService = new BasicUserService(userRepository);
         ChannelService channelService = new BasicChannelService(channelRepository);
         MessageService messageService = new BasicMessageService(messageRepository, channelRepository, userRepository);
+*/
+        // 스프링 컨텍스트 생성
+        ApplicationContext context = new AnnotationConfigApplicationContext(DiscodeitApplication.class);
 
-        // 셋업
+        // 필요한 Bean 가져오기
+        UserService userService = context.getBean(UserService.class);
+        ChannelService channelService = context.getBean(ChannelService.class);
+        MessageService messageService = context.getBean(MessageService.class);
+
+        // DiscodeitApplication에 옮겨둔 static 메소드 호출
+        User user = DiscodeitApplication.setupUser(userService);
+        Channel channel = DiscodeitApplication.setupChannel(channelService);
+        DiscodeitApplication.messageCreateTest(messageService, channel, user);
+
+        /*// 셋업
         User user = setupUser(userService);
         Channel channel = setupChannel(channelService);
         // 테스트
-        messageCreateTest(messageService, channel, user);
+        messageCreateTest(messageService, channel, user);*/
     }
 }
