+++ src/main/resources/application.yaml
@@ -1,0 +1,5 @@
+server:
+  port: 8081
+spring:
+  application:
+    name: MyApp
